% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/asosi.R
\name{asosi}
\alias{asosi}
\title{Title}
\usage{
asosi(date, region = c("casey", "davis", "durville", "mawson", "shackleton",
  "terranova", "westice", "ragnhild", "enderby", "capeadare", "sabrina"),
  band = c("infrared", "visible"))
}
\arguments{
\item{date}{}

\item{region}{}

\item{band}{}
}
\description{
Title
}
\examples{
\dontrun{
dates <- Sys.Date() - c(1, 2, 3, 4, 5)
for (i in seq_along(dates)) {
r <- asosi(dates[i])
writeRaster(r, sprintf("infrared\%s.tif", format(dates[i])))
r2 <- asosi(dates[i], band = "visible")
writeRaster(r2, sprintf("visible\%s.tif", format(dates[i])))
}
## prepare an object to build graticule lines
temp <- as(extent(r), "SpatialPolygons")
#' projection(temp) <- projection(r)

plot(r);llgridlines(temp)
}
}

